Extract certification-result-service
